fix vari

+++ client/src/components/Authentication/DeleteAccount.js
@@ -9,9 +9,8 @@ function DeleteAccount() {
 
     const [confirmEmail, setConfirmEmail] = useState("");
     const [error, setError] = useState("");
-    const [successMessage, setSuccessMessage] = useState("");
     const {showNotification} = useNotification();
-    const {user, sleep, logout, deleteAccount} = useAuth();
+    const {sleep, logout} = useAuth();
 
     const navigate = useNavigate();
 
@@ -19,23 +18,13 @@ function DeleteAccount() {
         event.preventDefault()
         setError("");
         try {
-            const trueEmail = user.email //trueEmail è l'effettiva mail dell'utente e la confronto con quella appena inserita
-            if (trueEmail !== confirmEmail) {
-                throw new Error("L'email non corrisponde a quella del tuo account. Riprova");
-            }
-            //se l'email corrisponde, si procede ad eliminare l'utente (tramite chiamata API al server)
-            await api.delete('http://localhost:5001/api/auth/delete-account')
-            setSuccessMessage("Eliminazione dell'account avvenuta correttamente!")
+            await api.delete(`http://localhost:5001/api/auth/delete-account/${confirmEmail}`)
+            showNotification("Eliminazione dell'account avvenuta correttamente!", "success");
             await sleep(2000);
             logout();
             navigate("/");
         }catch(error){
-            if(error.response){ //errore generato dall'endpoint API
-                setError(error.message);
-            }else{
-                setError(error.message); //errore dovuto ad email non corrispondenti
-            }
-            setConfirmEmail("");
+            showNotification(error.response.data, "error");
         }
     }
 
@@ -56,7 +45,6 @@ function DeleteAccount() {
                 <p></p>
                 <p></p>
                 <Typography component="p">Se non desideri più eliminare il tuo account, clicca <NavLink to="/">qui</NavLink> per tornare alla home</Typography>
-                {successMessage && <Typography component="p" className="success-message">{successMessage}</Typography>}
             </Box>
         </Box>
